fixing valgrind leaks --> test evaluation sheet
